+++ frontend/src/pages/budgetLines/DesiredAwardDate.jsx
@@ -80,3 +80,5 @@ export const DesiredAwardDate = () => {
         </div>
     );
 };
+
+export default DesiredAwardDate;
